+++ src/joedb/journal/Memory_File.cpp
@@ -8,7 +8,7 @@ namespace joedb
  size_t Memory_File::pread(char *buffer, size_t size, int64_t offset)
  ////////////////////////////////////////////////////////////////////////////
  {
-  const size_t max_size = data.size() - offset;
+  const size_t max_size = data.size() - size_t(offset);
   const size_t n = std::min(size, max_size);
   std::copy_n(data.data() + offset, n, buffer);
   return n;
@@ -18,10 +18,10 @@ namespace joedb
  void Memory_File::pwrite(const char *buffer, size_t size, int64_t offset)
  ////////////////////////////////////////////////////////////////////////////
  {
-  const size_t end = offset + size;
+  const size_t end = size_t(offset) + size;
   if (end > data.size())
    data.resize(end);
-  std::copy_n(buffer, size, &data[offset]);
+  std::copy_n(buffer, size, &data[size_t(offset)]);
  }
 
  ////////////////////////////////////////////////////////////////////////////
